Final adjustment and UI glitch fixed
Made changes so that the invalid command message will appear. Changed the stream insertion operator to have less UI elemements and focus more on its functionality. The state that the engine begins with was Win, but now it has been fixed to start at Start state. Removed additional files for State. They will remain in the Game Engine for convenience and the Game Engine is the only class that uses and has access to it