Adjustment and fixed since the last commit
When executing and removing an order, the program no longer crashes. No more issues with territories and their pointer to the player that owns it. Some small adjustment to make UI more clear. Listof territories to attack now displays territories player doesn't own, but should be changed to ones that adjacent to ones they own. Bonus armies have not been implemented yet. Performing an advance has an issue where it doesn't move your armies to teritory you own because it doesn't consider it as your own if you don't have armies on it.