Implemented Command Processor and Win State
When a command would be entered by the user in order to change states, the command processor will now take in the input as a command and store it. Win state has also been fully implemented where, once the game is over and reaches the win state, a user can decide whether to end or replay.